#Synchronization. Added chat with the team. Fixed bugs with registration.
