+++ game/mechanics/sendMessage.go
@@ -1,4 +1,4 @@
-package game
+package mechanics
 
 import (
 	"fmt"
@@ -12,9 +12,6 @@ func SendMessageFormChat(bot *tgbotapi.BotAPI, clan betypes.Clan, update tgbotap
 	go func() {
 		if update.Message.Text != " " {
 			text := ""
-			//getMessage := update.Message.Text
-			//getUserName := update.Message.From.UserName
-			//getFirstName := update.Message.From.FirstName
 
 			if len(users) == 0 {
 				for _, user := range clan.Users {
